Disable in-memory caching of scans during the load in certain circumstances
Only applies if disk caching is enabled (the default)
If we expect to read more scans than the in-memory pool size, then skip using the in-memory pool during the file load.
This reduces the amount of memory used during load, as well as reducing some overhead processing during the load as well.