Fix a data-replacement issue/bug in clsSpectraCache
Because it doesn't implement a least-recently-used cache and replaces data, it is possible to retrieve one spectrum from the in-memory pool to compare to a second one that must be read from disk, and end up with mangled data because the read from the disk replaced the list contents of the spectrum retrieved from the in-memory pool.
This change avoids the issue by always using a new object when reading from the disk.
Issue was confirmed by completely disabling the use of the in-memory pool.